requirements.txt
# FastAPI y dependencias web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
httpx>=0.24.0,<0.25.0

# Supabase y base de datos
supabase>=2.0.3,<2.1.0
postgrest>=0.10.8

# OpenAI y LangChain
openai==1.6.1
langchain==0.1.0
langchain-core==0.1.9
langchain-community==0.0.10
langchain-openai==0.0.2

# Utilidades
python-dotenv==1.0.0
pydantic>=2.5.2
pydantic[email]>=2.0.0
tiktoken>=0.5.1

# Dependencias adicionales para producción
gunicorn==21.2.0
uvloop==0.19.0
httptools==0.6.1
